% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_bar1}
\alias{plot_bar1}
\title{plot functions in MEPHAS}
\usage{
plot_bar1(count)
}
\arguments{
\item{count}{input count data frame}
}
\value{
bar plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
