% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_box2}
\alias{plot_box2}
\title{plot functions in MEPHAS}
\usage{
plot_box2(data2)
}
\arguments{
\item{data2}{input data frame}
}
\value{
box plot for 2 groups in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
