% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_boxm}
\alias{plot_boxm}
\title{plot functions in MEPHAS}
\usage{
plot_boxm(datam)
}
\arguments{
\item{datam}{input data frame}
}
\value{
box plot for more groups in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
