% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_coxstep}
\alias{plot_coxstep}
\title{plot functions in MEPHAS}
\usage{
plot_coxstep(data)
}
\arguments{
\item{data}{input data frame}
}
\value{
cpx-snell plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
