% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_density2}
\alias{plot_density2}
\title{plot functions in MEPHAS}
\usage{
plot_density2(data2)
}
\arguments{
\item{data2}{input data frame}
}
\value{
2 density plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
