% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_devres}
\alias{plot_devres}
\title{plot functions in MEPHAS}
\usage{
plot_devres(data)
}
\arguments{
\item{data}{input data frame}
}
\value{
deviance residuals plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
