% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_line1}
\alias{plot_line1}
\title{plot functions in MEPHAS}
\usage{
plot_line1(data, varx, vary)
}
\arguments{
\item{data}{input data frame}

\item{varx}{input x variable name}

\item{vary}{input y variable name}
}
\value{
line plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
