% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_mat}
\alias{plot_mat}
\title{plot functions in MEPHAS}
\usage{
plot_mat(data, ybreak)
}
\arguments{
\item{data}{input data frame}

\item{ybreak}{input break}
}
\value{
mat plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
