% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_msd2}
\alias{plot_msd2}
\title{plot functions in MEPHAS}
\usage{
plot_msd2(data)
}
\arguments{
\item{data}{input data frame}
}
\value{
mean sd for 2 groups plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
