% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_qq1}
\alias{plot_qq1}
\title{plot functions in MEPHAS}
\usage{
plot_qq1(data, varx)
}
\arguments{
\item{data}{input data frame}

\item{varx}{input x sample name}
}
\value{
qq plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
