% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plot_slgt}
\alias{plot_slgt}
\title{plot functions in MEPHAS}
\usage{
plot_slgt(data, varx, vary, xlab, ylab)
}
\arguments{
\item{data}{input data frame}

\item{varx}{input x variable name}

\item{vary}{input y variable name}

\item{xlab}{input x variable name}

\item{ylab}{input y variable name}
}
\value{
logistic plot in mephas
}
\description{
plot functions in MEPHAS
}
\examples{
# NOT Run

}
