% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{tabPanel.upload}
\alias{tabPanel.upload}
\title{Panel functions for mephas}
\usage{
tabPanel.upload(
  file = "file",
  header = "header",
  col = "col",
  sep = "sep",
  quote = "quote"
)
}
\arguments{
\item{file}{input data}

\item{header}{input header name}

\item{col}{input column  name}

\item{sep}{input separator name}

\item{quote}{input quote name}
}
\value{
panel for upload files used in mephas
}
\description{
Panel functions for mephas
}
\examples{
# NOT Run
  
}
